Alteracoes de layout e adicao de login guard

+++ API/SchoolNotyfyAPI_NEW/SchoolNotify.Domain/Entities/Responsavel.cs
@@ -11,7 +11,7 @@ namespace SchoolNotify.Domain.Entities
         public int Id { get; set; }
         public string Nome { get; set; }
         public string Email { get; set; }
-        public int Telefone { get; set; }
+        public long Telefone { get; set; }
         public virtual ICollection<Aluno> Alunos { get; set; }
     }
 }
